+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Rewrite core functionality by using cached state instead of relying on the file system.
   WARNING! this version is completely different functionally from the older one and the tags have to be recreated
   after updating it. [#30](https://github.com/wojciechkepka/wutag/pull/30)
+* Add `clean-cache` subcommand that cleans the cached tag registry from the filesystem.
 
 
 # 0.3.0